add custom string helper header
